@@ -12,7 +12,7 @@
 		}
 	],
 	"require": {
-		"php": ">=5.3.0",
+		"php": "^8.2",
 		"ext-pcntl": "*",
 		"colinmollenhour/credis": "1.2.*",
 		"psr/log": "^1.0.0 | ^2.0.0 | ^3.0.0"
