@@ -22,7 +22,7 @@ class Resque_Log extends Psr\Log\AbstractLogger
 	 * @param array   $context  Variables to replace { placeholder }
 	 * @return null
 	 */
-	public function log($level, $message, array $context = array())
+	public function log($level, string|\Stringable $message, array $context = []): void
 	{
 		if ($this->verbose) {
 			fwrite(
